+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [new version][unreleased]
 
+## Changes
+
+* Fixed flow for `contextmenu` event. Not it will be triggered on shapes too
+
 ## [2.0.2][2018-03-15]
 
 ## Fixed